+++ trabalho-pratico/src/model/user.c
@@ -144,7 +144,7 @@ void add_user_ride_data(User u, Ride r)
 {
     u->sum_score += get_ride_score_user(r);
     u->total_spent += get_ride_cost(r) + get_ride_tip(r);
-    u->total_distance += get_ride_distance(r);
+    u->total_distance += (unsigned short)get_ride_distance(r);
     unsigned short ride_date = get_ride_date(r);
     if (ride_date > u->last_ride_date) {
         u->last_ride_date = ride_date;
